backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
docx2txt
groq
python-dotenv
